Windows: Commonise logging code
