+++ Source/Windows/Common/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library(CommonWindows STATIC CPUFeatures.cpp InvalidationTracker.cpp)
+add_library(CommonWindows STATIC CPUFeatures.cpp InvalidationTracker.cpp Logging.cpp)
 target_link_libraries(CommonWindows FEXCore_Base)
 
 target_include_directories(CommonWindows PRIVATE
