+++ Source/Windows/Common/Logging.h
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: MIT
+namespace FEX::Windows::Logging {
+void Init();
+} // namespace FEX::Windows::Logging
